+++ packages/logstory/src/LoggerTypes.ts
@@ -1,3 +1,8 @@
 export type LogLevel = 'debug' | 'log' | 'warn' | 'error';
 export type Logger = Record<LogLevel, (...args: unknown[]) => void>;
 export type LogLevelState = Record<LogLevel, boolean>;
+export interface CreateLoggerParams {
+    name?: string;
+    logLevelState?: Partial<LogLevelState>;
+    consoleProxy?: Console;
+}
