+++ packages/logstory/src/index.ts
@@ -1,37 +1,13 @@
-import { Logger, LogLevel, LogLevelState } from './LoggerTypes';
+import { applyConsoleProxyMiddleware, MiddlewareFunction } from './utils/applyConsoleProxyMiddleware';
+import { createLogstory } from './createLogStory';
+import { Logger, LogLevel, LogLevelState, CreateLoggerParams } from './LoggerTypes';
 
-const defaultLogLevelState: LogLevelState = {
-    debug: true,
-    log: true,
-    warn: true,
-    error: true
-};
-
-interface CreateLoggerParams {
-    name?: string;
-    logLevelState?: Partial<LogLevelState>;
-    consoleProxy?: Console;
-}
-
-export const createLogstory = ({
-    name,
-    logLevelState = defaultLogLevelState,
-    consoleProxy = window.console
-}: CreateLoggerParams = {}): Logger => {
-    const loggerName = name ? `[${name}]: ` : '';
-    // https://stackoverflow.com/a/42406361/2973464
-    const logger: Logger = {
-        debug: consoleProxy.debug.bind(window.console, `🥷${loggerName}%o`),
-        log: consoleProxy.log.bind(window.console, `${loggerName}%o`),
-        warn: consoleProxy.warn.bind(window.console, `😕😕😕️\n${loggerName}%o`),
-        error: consoleProxy.error.bind(window.console, `🔥🔥🔥🔥🔥🔥\n${loggerName}%o`)
-    };
-
-    (Object.keys(logger) as Array<LogLevel>).forEach((key) => {
-        if (!logLevelState[key]) {
-            logger[key] = () => {};
-        }
-    });
-
-    return logger;
+export {
+    createLogstory,
+    applyConsoleProxyMiddleware,
+    MiddlewareFunction,
+    Logger,
+    LogLevel,
+    LogLevelState,
+    CreateLoggerParams
 };
